Add vc only scripts to extras
add vehicle controller only usage speedometer scripts to extras folder as those would be needed for vehicle controller only usage and. then opting for speedometer usage, as current speedometer housed in repo is further edited for vehicle entry usages and thus adding the alt for controller only usage in extras as an optional was needed.

Still need to write docs for these as well as all other documentations writings if and or when my further free time allows.