Update AllVehicleControllers-wEntry-Documentation.txt
Continue writing: AllVehicleControllers-wEntry-Documentation.txt

---- WRITING OF THIS DOCUMENTATION IS STILL IN PROGRESS!!!! ----

Note: As this particular doc will be rather lengthly  we will add to it as I complete further parts, but, for now I will not add the linkage to it in the readme as such is not yet completed. Hopefully that makes sense. :)

Parts Currently Completed:

Ambulance 01 - 04
MuscleCar 01 - 04
PoliceCruiser 01 - 04
SedanLarge 01 - 04
SedanMedium 01 - 04
SedanSmall 01 - 04
Taxi 01 - 04